core/requirements.txt
SQLAlchemy>=1.3.7
SQLAlchemy-Utils>=0.36.8
psycopg2-binary>=2.8.5
aiofiles>=0.5.0
certifi>=2020.4.5.1
chardet>=3.0.4
Click>=7.0
colorama>=0.4.3
fastapi>=0.65.2
h11>=0.9.0
httptools>=0.0.13
idna>=2.8
numpy>=1.19.2
prompt-toolkit>=2.0.10
pydantic>=1.6.2
Pygments>=2.3.1
python-dateutil>=2.8.1
python-multipart>=0.0.5
pytz>=2020.1
PyYAML>=5.3.1
requests>=2.22.0
six>=1.15.0
starlette>=0.13.2
urllib3>=1.24.3
uvicorn>=0.11.2
uvloop>=0.14.0
wcwidth>=0.1.9
websockets>=8.1
PyJWT>=2.1.0
passlib>=1.7.4
bcrypt>=3.2.0
pg8000==1.23.0
pgvector
llama-cpp-python